requirements.txt
streamlit==1.27.0
tensorflow==2.13.0
Pillow==9.3.0
numpy==1.24.3
